requirements.txt
django
django-filter
djangorestframework
drf-nested-routers
django-rest-knox
Pillow
cryptography
easy-thumbnails
gunicorn
